refactor: authファイルを責務別プレフィックスで整理
- credentials・session・oauthごとに型とインターフェースを集約
- OAuthプロバイダーとセッションリポジトリのファイル名を統一
- 関連テストを本体と同じプレフィックスへ変更
