+++ internal/feature/auth/credentials_email_test.go
