+++ internal/feature/auth/session.go
@@ -10,7 +10,44 @@ import (
 	"time"
 )
 
-const refreshTokenBytes = 32
+const (
+	// DefaultRefreshTokenTTL はリフレッシュトークンの有効期限です。
+	DefaultRefreshTokenTTL = 30 * 24 * time.Hour
+
+	// refreshTokenReuseGracePeriod は正規クライアントの並行更新を盗難と誤検知しないための猶予期間です。
+	refreshTokenReuseGracePeriod = 30 * time.Second
+
+	refreshTokenBytes = 32
+)
+
+// TokenPair はブラウザセッションへ発行するアクセストークンと
+// リフレッシュトークンをまとめた値です。
+type TokenPair struct {
+	AccessToken      string
+	RefreshToken     string
+	RefreshExpiresAt time.Time
+}
+
+// RefreshSession はサーバー管理のリフレッシュセッションです。
+// TokenHash にはトークン本体ではなく SHA-256 ハッシュだけを保持します。
+type RefreshSession struct {
+	ID         string
+	FamilyID   string
+	UserID     int64
+	TokenHash  []byte
+	ExpiresAt  time.Time
+	ConsumedAt *time.Time
+	RevokedAt  *time.Time
+	ReplacedBy *string
+	CreatedAt  time.Time
+}
+
+// JWTGenerator はJWTトークン生成のインターフェースを定義します。
+// Goの慣例に従い、インターフェースはプロバイダー（platform/jwt）ではなくコンシューマー（usecase）が定義します。
+type JWTGenerator interface {
+	// GenerateToken は指定されたユーザーの署名済みJWTトークンを生成します。
+	GenerateToken(userID int64, email string) (string, error)
+}
 
 // RefreshSessionRepository はリフレッシュセッションの永続化操作を定義します。
 type RefreshSessionRepository interface {
